completes budget increase decrease

+++ db/money_tracker.sql
@@ -6,26 +6,26 @@ DROP TABLE IF EXISTS merchants;
 
 CREATE TABLE budgets(
   id SERIAL8 PRIMARY KEY,
-  amount MONEY
+  amount INT8
 );
 
 CREATE TABLE categories(
   id SERIAL8 PRIMARY KEY,
   name VARCHAR(255),
-  total_spend MONEY
+  total_spend INT8
 );
 
 CREATE TABLE merchants(
   id SERIAL8 PRIMARY KEY,
   name VARCHAR(255),
-  total_spend MONEY
+  total_spend INT8
 );
 
 
 CREATE TABLE transactions(
   id SERIAL8 PRIMARY KEY,
   item VARCHAR(255),
-  price MONEY,
+  price INT8,
   merchant_id INT8 REFERENCES merchants(id) ON DELETE CASCADE,
   category_id INT8 REFERENCES categories(id) ON DELETE CASCADE,
   date VARCHAR(255)
